Initial Commit of Everything Functioning
