+++ ir.py
@@ -1,66 +1,44 @@
 import RPi.GPIO as GPIO
 import time
-import threading
 
 class IRRemoteHandler:
-    def __init__(self, receiver_pin, speech_engine, scanner_ui):
+    def __init__(self, receiver_pin):
         """Initialize IR remote handler and setup GPIO."""
         self.receiver_pin = receiver_pin
-        self.speech = speech_engine
-        self.scanner = scanner_ui  # Store ScannerUI reference
-        self.callbacks = {}  # Stores button mappings to functions
         self.setup_gpio()
-        self.setup_callbacks()  # Move all function mappings inside the class
-   
-    def listen(self):
-        """Loop to listen for IR signals (Replace with actual IR input handling)"""
-        print("[INFO] Listening for IR remote inputs...")
-        while True:
-            # Simulated input handling (Replace with GPIO input logic)
-            time.sleep(1)
-    
+
     def setup_gpio(self):
         """Set up GPIO for the IR receiver."""
         GPIO.setmode(GPIO.BCM)
         GPIO.setup(self.receiver_pin, GPIO.IN)
 
-    def setup_callbacks(self):
-        """Maps IR remote buttons to their corresponding actions."""
-        self.callbacks = {
-            "KEY_UP": (self.scanner.channel_up, "Channel Up"),
-            "KEY_DOWN": (self.scanner.channel_down, "Channel Down"),
-            "KEY_LEFT": (self.scanner.zone_down, "Zone Down"),
-            "KEY_RIGHT": (self.scanner.zone_up, "Zone Up"),
-            "KEY_HASH": (self.volume_down, "Volume Down"),
-            "KEY_STAR": (self.volume_up, "Volume Up"),
-        }
-
     def listen(self):
-        """Continuously listens for IR signals."""
+        """Continuously listens for IR signals and prints the key pressed."""
         print("Listening for IR remote inputs...")
-        while True:
-            if GPIO.input(self.receiver_pin) == GPIO.LOW:
-                code = self.decode_signal()
-                if code and code in self.callbacks:
-                    self.speech.speak(self.callbacks[code][1])  # Speak button function
-                    self.callbacks[code][0]()  # Execute assigned function
-                time.sleep(0.1)
+        try:
+            while True:
+                if GPIO.input(self.receiver_pin) == GPIO.LOW:
+                    code = self.decode_signal()
+                    if code:
+                        print(f"Key Pressed: {code}")  # Print the key name
+                    time.sleep(0.1)
+        except KeyboardInterrupt:
+            print("\nExiting...")
+            self.cleanup()
 
     def decode_signal(self):
-        """Placeholder for decoding IR signals (to be implemented)."""
-        # Implement decoding logic here (returning button codes)
-        pass
-
-    def volume_up(self):
-        """Increases volume and speaks confirmation."""
-        # Placeholder: Add logic to increase volume
-        self.speech.speak("Volume increased")
-
-    def volume_down(self):
-        """Decreases volume and speaks confirmation."""
-        # Placeholder: Add logic to decrease volume
-        self.speech.speak("Volume decreased")
+        """Placeholder for IR decoding logic. Simulates button presses."""
+        # In actual implementation, replace this with IR decoding logic.
+        # For now, return a simulated button name.
+        return "KEY_UP"  # Simulated key press
 
     def cleanup(self):
         """Cleanup GPIO on exit."""
         GPIO.cleanup()
+        print("GPIO cleaned up.")
+
+# Run the script directly
+# if __name__ == "__main__":
+#     receiver_pin = 18  # Change to the correct GPIO pin
+#     remote = IRRemoteHandler(receiver_pin)
+#     remote.listen()
