+++ logs/app_log.txt
@@ -419,3 +419,37 @@
 [2025-03-16 10:47:51] [INFO] Connection Status: OP25 Connected
 [2025-03-16 10:51:35] [INFO] Connection Status: OP25 Connected
 [2025-03-16 13:01:56] [INFO] Connection Status: OP25 Connected
+[2025-03-16 13:11:27] [INFO] Connection Status: OP25 Connected
+[2025-03-16 13:14:21] [INFO] Connection Status: OP25 Connected
+[2025-03-16 13:36:14] [INFO] Connection Status: OP25 Connected
+[2025-03-16 13:41:23] [INFO] Connection Status: OP25 Connected
+[2025-03-16 13:46:13] [INFO] Connection Status: OP25 Connected
+[2025-03-16 13:47:34] [INFO] Connection Status: OP25 Connected
+[2025-03-16 13:52:35] [INFO] Connection Status: OP25 Connected
+[2025-03-16 13:53:02] [INFO] Connection Status: OP25 Connected
+[2025-03-16 13:57:51] [INFO] Connection Status: OP25 Connected
+[2025-03-16 13:58:29] [INFO] Connection Status: OP25 Connected
+[2025-03-16 13:59:26] [INFO] Connection Status: OP25 Connected
+[2025-03-16 13:59:55] [INFO] Connection Status: OP25 Connected
+[2025-03-16 14:00:58] [INFO] Connection Status: OP25 Connected
+[2025-03-16 14:04:27] [INFO] Connection Status: OP25 Connected
+[2025-03-16 14:07:08] [INFO] Connection Status: OP25 Connected
+[2025-03-16 14:08:24] [INFO] Connection Status: OP25 Connected
+[2025-03-16 14:09:34] [INFO] Connection Status: OP25 Connected
+[2025-03-16 14:10:42] [INFO] Connection Status: OP25 Connected
+[2025-03-16 14:11:15] [INFO] Connection Status: OP25 Connected
+[2025-03-16 14:15:33] [INFO] Connection Status: OP25 Connected
+[2025-03-16 14:15:56] [INFO] Connection Status: OP25 Connected
+[2025-03-16 14:16:24] [INFO] Connection Status: OP25 Connected
+[2025-03-16 14:16:49] [INFO] Connection Status: OP25 Connected
+[2025-03-16 14:17:25] [INFO] Connection Status: OP25 Connected
+[2025-03-16 14:18:11] [INFO] Connection Status: OP25 Connected
+[2025-03-16 14:37:02] [INFO] Connection Status: OP25 Connected
+[2025-03-16 14:38:50] [INFO] Connection Status: OP25 Connected
+[2025-03-16 14:54:07] [INFO] Connection Status: OP25 Connected
+[2025-03-16 15:04:18] [INFO] Connection Status: OP25 Connected
+[2025-03-16 15:05:09] [INFO] Connection Status: OP25 Connected
+[2025-03-16 15:06:39] [INFO] Connection Status: OP25 Connected
+[2025-03-16 15:11:09] [INFO] Connection Status: OP25 Connected
+[2025-03-16 17:03:10] [INFO] Connection Status: OP25 Connected
+[2025-03-16 17:06:30] [INFO] Connection Status: OP25 Connected
